Added a Location Context Provider.
You can now obtain the the user's prefferred
Country and State information without needing
to query the backend every time you need it them.

That being said, the backend will still store such data
in redis, so that the data is not lost between user sessions
or sign outs/in. The data is altered and PATCHed to the backend
in the Accounts page. The data is retrieved whenever the user changes
their account (i.e. registers or signs in/out).

The specific location data stored is the countryCode/_id shared
between the iso2 and countries json files, respectively. The States
are stored as 'state' and are the abbreviations used in the usStates
json.