Overhauled the CasesTracking file, and some other minor changes.
Tried to implement the Location Context Provider in my original
Cases Tracking file, but due to the way I originally coded it
(which was an overly complicated way to get around essentially
three different axios calls to different urls), I had to
essentially purge half of it, and implement the removed functionality
elsewhere in the code. Implementing the location Context Provider
in the other files will likely be much easier than this one
(which was just complicated due to my convuluted first draft), and
I'll implement that tomorrow, as well as more caching for api queries
using redis.

The other changes were fairly minor and (probably) won't cause any
major bugs. That being said, I haven't extensively tested it yet,
and will do so tomorrow morning.

It's three AM, and I'm going to sleep.