Implemented Location Context for Travel
Order of jsons and lists is EXTREMELY important
to this implementation. DO NOT CHANGE the order.

If you absolutely must change the order, please consult
Loughlin Claus before doing so.

Travel component now defaults to the user's location
preference when possible. However, this will not always
be possible. If their specified country is not contained
within either geoData.json or countries-iso2.json,
or they are simply not signed in, it will default back
to USA (the original default). Further notes on implementation
can be found in the comments of the findCountry function.