Frontend HospitalBeds & Vaccine now query backend.
Implemented these two quickly, as both just use a static route for
literaly every query. I also added a try/catch to HospitalBeds, for the
(unlikely) scenario where there is an error in the axios call.

There is a bit of a delay in the Hospital Capacity loading time. This is
a result of the sheer size of the original API query's response (over
200,000 lines), and was originally an unavoidable problem. However, now
that I've implemented redis api routing on the backend, I can store each
individual state's data in the server's redis cache.

This would decrease the load time of the Hospital Capacity page by
nearly 98%. So I'll look into implementing that after I finish the rest
of the frontend.

Though it would increase the load time of the first time the
page is visited by a tad. But in the long run, I think it's worth it.