Frontend Testing, Travel, & eachTesting query back.
Implemented the backend redis api queries for the frontend's Testing,
Travel, and eachTestingLocation components.

I will now work on merging the auth branch into the development branch.

Sorry about the delay on this commit. In all honesty, it took less then
10 minutes to code the whole thing. But halfway through, I got a bit
destracted by the efficiency of the code, so I ran some tests between
the backend redis api load times and the original api loadtimes.

While the backend redis api loads pages 10x faster if a page data has
been previously cached, it is about 3x slower on the first, uncached visit to
a page. I've thought of a solution, but it would take some time to
implement, so I'm going to do the merge to development first, then
work to decrease the uncached webpage loadtimes. Then I'll do the
eachhospital be thing I thought of in the previous commit.