Merged branches v2.1
This is the result of the intial merge attempt of the auth
branch and the development branch.

The only page that seems to be genuinely damaged by the merge is the
Account page. The navbar also seems to be missing a fix I made for
when a newline occured that prevented the lines from overlapping.

Still, if possible, I'd like for everyone to at least glance over any
alterations I made to their code before I push it to the development
branch. Until then, I'll continue working to fix problems caused by the
merge.