Fixed security vulnerability in news.js.
Turns out, some of the available Countries returned empty articles
arrays when querying the news API. Normally, I wouldn't care much, and
just have the frontend display a "No data found for country" message &
be done with it. However, because the API returned a completely
syntactically valid response (just an empty array for the 'articles'
key/value pair), the code runs without crashing, but no news data would
be cached for the country. Without news data for a country being cached,
client.exists would return false, and any repeated requests for that
country would have the backend requery the news API url. This could
theoretically allow a single client to request news data for a single
(news-less) country 50 times in a single day, preventing all other
clients from seeing any news for that day (kind of like a simple,
single-source DDoS attack).

Fixed it by taking all the prior 'isValid' country checks outside the
first else-if (that decides whether to retreive world data or country
data) and adding one 'valid country param' check that, after passing all
prior checks, views whether the cache has actual news articles for the
country or if it contains an arbitrary 'we already checked this country
today, and there was no news data' value, & if it contains this
arbitrary value, it will default to world news.

And in the 'query api for new news country data' route, I have it check
if the query returned any articles, and if not, cache the arbitray 'no
data' value for that country.

Also, just to be safe, I removed the country that retuned an empty array
from the available Countries array (cn).

I also added package-lock.json to the .gitignore, because when I switch
branches I keep stupidly running npm install on the entire project, and
I don't want to accidentally push the resulting package-lock someday.