Increased time range of chart.
Now instead of showing only the Covid chart data of only the past month
or so, it shows all the data the api has on the country since 1/22/20.

This likely lessens the prior load time improvement of altChart, but
upon review, the load time differnces were practically negligible from
the start (maybe 100-200 ms). Honestly, the load time isn't much of a
problem for the Chart page in the first place, especially in comparison
to some other pages of the site (which I'll look into decreasing
tomorrow).

For notes on differences between the altChart branch and the development
branch, see yesterday's commit #cd2229d.

Its once again 3AM, and I'm going to bed.
Reeaalllllly need to fix my sleep schedule.