altChart v2
I consider this iteration of the altChart to be different enough from
the previous to warrant a new version number. The team MUST meet
sometime to discuss which version of the altChart to use (Maithili's
original, Maithili's modified, altChartv1 (undo this commit), or
altChartv2), as each group member seems to have their own opinions on
this matter.

For this commit note, I will first go into a brief overview of the
changes between altChartv1 and altChartv2. Then go into the technical
aspects of these changes. Lastly, I will be adding a detailed comment on
the technical aspects of the original altChart commit on that commit's
page. The technical parts can probably be ignored by most of the group,
they're more for Maithili (who I understand is making their own modified
version of the orginal chart), in case they want to reference/implement
any of the changes I made in the altCharts into their version.

Brief Overview of altChartv2:
-moves chart to the homepage (& if the chart can be rendered, it will
replace the statistics data table, as the chart contains all the data
(and more) that the table provides)
-removes the chart webpage
-adds commas to the numbers in the chart hover tool (For better
readability)
-the animation of the chart seems to have become a bit buggier than when
it had its own page (and loses that cool transition between data charts)

Technical Overview (small stuff first):
-App.css: set the chartDiv's margin to auto & width to 50% to better
center the chart on screens of varying sizes
-App.js & Navigation.js: removed the Chart Page links

-Charts.js:
--Added formatter to the tooltip, giving the hover data numbers commas
--Commented out all the data computation stuff (like the useEffect &
such), as now it's all done in the CasesTracking.js file & then passed
in using props

-CasesTracking.js (homepage):
--Added chartArr & chartDataFound params to the buildDisplay function,
and uses them to render the chart when possible (when chart data is
found & when it is not creating a state). When not possible, it displays
a no chart data found message (if not a state) and shows the original
statistics data table instead
--Removed/swapped out Iso3 conversion stuff, as I found out during
altChartv1 that I can just use the country _id in queries instead
(making the code imperceptibly faster, and easier to combine with the
Charts.js code, which uses _id for queries instead of iso3's).
--brought over the data computation stuff from the Chart.js file, so
that all queries are done in the same use Effect.
--removed getChartData's recursive call for all world data when country
data is not found, as page will now show statistical data table in that
scenario.