More efficient hospitalBed queries.
I've decreased the query/load time for cached hospital beds by
about 80-90% (from 373ms => 40-70ms). I also tried implementing a
potential efficiency reduction for all the 'first visit' page loads, but
the time difference for that was fairly negligible, so I scrapped it.

Unfortunately, this means that uncached data page loads can still take a
long-ass time. In some cases, several seconds long. At this point, the
only possible solution to that 'problem' is if we cached the larger data
sets on server start-up. But that's only a stop-gap, and not really a
great solution for the long-term.

I left the previous hospital bed query in the code, all you need to do
to switch back to it is remove /${state} from line 21 of the frontend's
HospitalBeds.js.

I decrease the initial, uncached loadtime of the hospitalBeds page by
caching the data only after sending it, but honestly redis works so fast
that it doesn't make much of a difference in to the overall loadtime
(especially in light of what I think is the main problem, that the api
sends over 200,000 lines in response to every query).

The way I went about the efficiency changes is by basically having redis
store bedUtilization for every individual state instead of all of them
together. Thus, the code only needs to parse through about 4,000 entries
insteadof checking (& performing string equivalence checks) on over
200,000 entries. The code now queries redis for a list of a state's
entries instead of all of them (and then searching response for
matches). Kind of repeated myself there.

I also added two other hospitalBeds' API-specific states to the end of
the dropdown, Puerto Rico & district of Columbia. Haven't tested them
though. So if they don't work, I'll remove them later.

We really need to have a group meeting with everyone sometime present
sometime soon.