Final Main Branch Commit.
Added error checking to the client side, now displays error messages
when the server is down, or if a query fails on any page that interacts
with the  backend. Also fixed a bug in first-party user registeration
where the user's name was not properly stored.

Also, I returned the News.js page's dropdown to the version with the 200
extra "impossible to get news, always default to Earth" options. I also
altered some comment.log's on the backend's news.js page.