Merged Branches v3.3: re-implement BarChart
I was being really dumb about the way I was re-implementing the
barchart. Then I realized that I don't have to re-query all the data
every time the selected country changes. Unlike the original, I'm
performing two separate api queries. And the data for the top 5
countries is relatively static in comparison to the repeated queries for
the user's selected countries.

I only need to get 1 days worth of data from all the countries of the
world once per page visit to make the code work. So I set it to only
send the request for the top 5 countries data set once per page visit.
Kind of obvious in hindsight.

I uncommented the Bar Chart code, and limited the bar charts queries to
only a single day (instead of a month or a year, unlike the other
versions). Other than that, the bar chart code is basically the same as
Maithili's OG version.

Gonna switch to the Auth branch and try the most difficult (and final)
merges with the UI code there.